openlineage-playground/scripts/pyproject.toml
[project]
name = "openlineage-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openlineage-python>=1.33.0",
    "openlineage-sql>=1.33.0",
    "rich>=14.0.0",
    "sqlglot>=26.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
